# Pad dimensions for faster 1D convolutions