Addresses #7961. We may want to try to fix this at the XLA:CPU level instead, but this is a quick fix from the Python side.

Here's a brief test script I ran via IPython:
```python
import jax.numpy as jnp
from jax import jit
from jax.scipy import signal
import numpy as np
import jax.numpy as jnp

x = jnp.ones(10000)
y = jnp.ones(100)

assert np.allclose(jit(jnp.convolve)(x, y), np.convolve(x, y))
assert np.allclose(jit(signal.convolve)(x, y), np.convolve(x, y))

print('np.convolve:               ', end='')
%timeit np.convolve(x, y)

print('jax.numpy.convolve:        ', end='')
%timeit jit(jnp.convolve)(x, y).block_until_ready()

print('jax.scipy.signal.convolve: ', end='')
%timeit jit(signal.convolve)(x, y).block_until_ready()
```
Tested on my local CPU...

Before this PR :
```
np.convolve:               237 µs ± 4.08 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
jax.numpy.convolve:        2.29 ms ± 54.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
jax.scipy.signal.convolve: 2.25 ms ± 22.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
After this PR:
```
np.convolve:               209 µs ± 2.16 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
jax.numpy.convolve:        261 µs ± 22.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
jax.scipy.signal.convolve: 259 µs ± 9.15 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```